Add code for initial test case
This grabs all traffic on wire, interprets the traffic
as IPv4, grabs the dest addr, gets the hostname, and
prints it.

Next is to figure out how to do inspection. Maybe we want to
try to rip the name out of the packet? Maybe we use a static
list of IP's from the targeted ad services, etc. Fun problems
to solve now...